+++ lib/widgets/settings/review.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:iconsax/iconsax.dart';
+import 'package:in_app_review/in_app_review.dart';
+
+Widget buildReviewSettings(BuildContext context) {
+  return Card(
+    clipBehavior: Clip.hardEdge,
+    margin: const EdgeInsets.all(8.0),
+    child: ListTile(
+      title: Text(AppLocalizations.of(context)!.review),
+      subtitle: Text(AppLocalizations.of(context)!.reviewSub),
+      leading: const Icon(Iconsax.like),
+      onTap: () async {
+        final InAppReview inAppReview = InAppReview.instance;
+
+        if (await inAppReview.isAvailable()) {
+          inAppReview.requestReview();
+        }
+      },
+    ),
+  );
+}